iox-#1560 Move filesystem-related methods from 'helplets.hpp' to 'filesystem.hpp'
