+++ iceoryx_hoofs/include/iceoryx_hoofs/internal/posix_wrapper/unix_domain_socket.hpp
@@ -17,6 +17,7 @@
 #ifndef IOX_HOOFS_POSIX_WRAPPER_UNIX_DOMAIN_SOCKET_HPP
 #define IOX_HOOFS_POSIX_WRAPPER_UNIX_DOMAIN_SOCKET_HPP
 
+#include "iceoryx_hoofs/cxx/filesystem.hpp"
 #include "iceoryx_hoofs/internal/posix_wrapper/ipc_channel.hpp"
 #include "iceoryx_hoofs/internal/units/duration.hpp"
 #include "iceoryx_platform/fcntl.hpp"
@@ -25,7 +26,6 @@
 #include "iceoryx_platform/stat.hpp"
 #include "iceoryx_platform/un.hpp"
 #include "iox/optional.hpp"
-#include "iceoryx_hoofs/cxx/helplets.hpp" // isValidFileName
 
 namespace iox
 {
